requirements.txt
streamlit
langchain
langchain-google-genai
langchain-openai
langchain-community
langgraph
supabase
python-dotenv
pypdf
tiktoken
plotly